changed IPlayer interface
