+++ tests/test/test_player.cpp
@@ -45,18 +45,18 @@ TEST(Players, SaveLoad)
 
     for (auto & player : players) {
 
-        player->Play(dc::Vector2(0.1f, 2.f));
+        player->Play({ dc::Vector2(0.1f, 2.f), dc::moves::Shot::Rotation::kCCW });
 
         auto storage = player->CreateStorage();
-        auto const shot0 = player->Play(dc::Vector2(0.1f, 2.f));
+        auto const shot0 = player->Play({ dc::Vector2(0.1f, 2.f), dc::moves::Shot::Rotation::kCCW });
 
         player->Load(*storage);
-        auto const shot1 = player->Play(dc::Vector2(0.1f, 2.f));
+        auto const shot1 = player->Play({ dc::Vector2(0.1f, 2.f), dc::moves::Shot::Rotation::kCCW });
 
         // PrintVector2(shot0);
         // PrintVector2(shot1);
 
-        EXPECT_TRUE(EqualsVector2(shot0, shot1)) << player->GetPlayerId();
+        EXPECT_TRUE(EqualsVector2(shot0.velocity, shot1.velocity)) << player->GetPlayerId();
     }
 }
 
@@ -67,18 +67,18 @@ TEST(Players, SaveLoad2)
     for (auto & player : players) {
         auto storage = player->CreateStorage();
 
-        player->Play(dc::Vector2(0.1f, 2.f));
+        player->Play({ dc::Vector2(0.1f, 2.f), dc::moves::Shot::Rotation::kCCW });
 
         player->Save(*storage);
-        auto const shot0 = player->Play(dc::Vector2(0.1f, 2.f));
+        auto const shot0 = player->Play({ dc::Vector2(0.1f, 2.f), dc::moves::Shot::Rotation::kCCW });
 
         player->Load(*storage);
-        auto const shot1 = player->Play(dc::Vector2(0.1f, 2.f));
+        auto const shot1 = player->Play({ dc::Vector2(0.1f, 2.f), dc::moves::Shot::Rotation::kCCW });
 
         // PrintVector2(shot0);
         // PrintVector2(shot1);
 
-        EXPECT_TRUE(EqualsVector2(shot0, shot1)) << player->GetPlayerId();
+        EXPECT_TRUE(EqualsVector2(shot0.velocity, shot1.velocity)) << player->GetPlayerId();
     }
 }
 
@@ -87,18 +87,18 @@ TEST(Players, SaveLoad3)
     auto players = CreatePlayers();
 
     for (auto & player : players) {
-        player->Play(dc::Vector2(0.1f, 2.f));
+        player->Play({ dc::Vector2(0.1f, 2.f), dc::moves::Shot::Rotation::kCCW });
 
         auto storage = player->CreateStorage();
-        auto const shot0 = player->Play(dc::Vector2(0.1f, 2.f));
+        auto const shot0 = player->Play({ dc::Vector2(0.1f, 2.f), dc::moves::Shot::Rotation::kCCW });
 
         auto player_copy = storage->CreatePlayer();
-        auto const shot1 = player_copy->Play(dc::Vector2(0.1f, 2.f));
+        auto const shot1 = player_copy->Play({ dc::Vector2(0.1f, 2.f), dc::moves::Shot::Rotation::kCCW });
 
         // PrintVector2(shot0);
         // PrintVector2(shot1);
 
-        EXPECT_TRUE(EqualsVector2(shot0, shot1)) << player->GetPlayerId();
+        EXPECT_TRUE(EqualsVector2(shot0.velocity, shot1.velocity)) << player->GetPlayerId();
     }
 }
 
@@ -107,19 +107,19 @@ TEST(Players, SaveLoad4)
     auto players = CreatePlayers();
 
     for (auto & player : players) {
-        player->Play(dc::Vector2(0.1f, 2.f));
+        player->Play({ dc::Vector2(0.1f, 2.f), dc::moves::Shot::Rotation::kCCW });
 
         auto storage = player->CreateStorage();
         json j_storage = *storage;
-        auto const shot0 = player->Play(dc::Vector2(0.1f, 2.f));
+        auto const shot0 = player->Play({ dc::Vector2(0.1f, 2.f), dc::moves::Shot::Rotation::kCCW });
 
         auto storage2 = j_storage.get<std::unique_ptr<dc::IPlayerStorage>>();
         auto player_copy = storage2->CreatePlayer();
-        auto const shot1 = player_copy->Play(dc::Vector2(0.1f, 2.f));
+        auto const shot1 = player_copy->Play({ dc::Vector2(0.1f, 2.f), dc::moves::Shot::Rotation::kCCW });
 
         // PrintVector2(shot0);
         // PrintVector2(shot1);
 
-        EXPECT_TRUE(EqualsVector2(shot0, shot1)) << player->GetPlayerId();
+        EXPECT_TRUE(EqualsVector2(shot0.velocity, shot1.velocity)) << player->GetPlayerId();
     }
 }
